[공개 버킷상세] 댓글 추가 api 에 int 타입을 사용하도록 수정

+++ app/src/main/java/com/zinc/berrybucket/ui/presentation/detail/screen/OpenDetailScreen.kt
@@ -382,7 +382,7 @@ fun OpenDetailScreen(
                                     is OpenDetailEditTextViewEvent.SendComment -> {
                                         viewModel.addBucketComment(
                                             request = AddBucketCommentRequest(
-                                                bucketlistId = detailId,
+                                                bucketlistId = detailId.toInt(),
                                                 content = it.sendText,
                                                 mentionIds = "" //TODO : 수정필요
                                             )
